# openapi.yaml
openapi: 3.0.3
info:
  title: Weather App with Favorites API
  version: 1.0.0
  description: |
    API for user authentication, weather lookup via WeatherAPI, and managing user favorites.

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Auth
  - name: Weather
  - name: Favorites

paths:
  /register:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: secret123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /login:
    post:
      tags:
        - Auth
      summary: Log in a user and receive a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: secret123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /weather:
    get:
      tags:
        - Weather
      summary: Get current weather for a city
      description: Fetches weather data from WeatherAPI through the backend only.
      parameters:
        - name: city
          in: query
          required: true
          description: City name to look up
          schema:
            type: string
          example: Vilnius
        - name: country
          in: query
          required: false
          description: Optional country name to narrow the search
          schema:
            type: string
          example: Lithuania
      responses:
        '200':
          description: Weather data returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherResponse'
        '400':
          description: Missing city parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: City not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Weather API failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server misconfiguration or internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /favorites:
    get:
      tags:
        - Favorites
      summary: List the authenticated user's favorites
      security:
        - bearerAuth: []
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token in the format `Bearer <token>`
          schema:
            type: string
          example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Favorites returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Favorite'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Favorites
      summary: Add a city to favorites
      security:
        - bearerAuth: []
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token in the format `Bearer <token>`
          schema:
            type: string
          example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteCreateRequest'
            example:
              city: Vilnius
      responses:
        '201':
          description: Favorite saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /favorites/{id}:
    delete:
      tags:
        - Favorites
      summary: Delete one favorite by id
      security:
        - bearerAuth: []
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token in the format `Bearer <token>`
          schema:
            type: string
          example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        - name: id
          in: path
          required: true
          description: Favorite id
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Favorite deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Favorite deleted
        '400':
          description: Invalid favorite id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Favorite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    RegisterResponse:
      type: object
      properties:
        message:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            email:
              type: string
              format: email

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token

    WeatherResponse:
      type: object
      properties:
        location:
          type: object
          properties:
            name:
              type: string
            region:
              type: string
            country:
              type: string
            localtime:
              type: string
        weather:
          type: object
          properties:
            temperatureC:
              type: number
            temperatureF:
              type: number
            condition:
              type: string
            icon:
              type: string
            humidity:
              type: integer
            windKph:
              type: number

    FavoriteCreateRequest:
      type: object
      required:
        - city
      properties:
        city:
          type: string
          minLength: 2

    Favorite:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        city:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string